Move startup function out of main urls
The `startup` function queries models when the main `urls` module is
imported which produces this error when a database is being created
because the necessary tables are not created yet:

```
django.db.utils.ProgrammingError: (1146, "Table 'SS.locations_space' doesn't exist")
```

This moves the function to its own module and calls it from the `wsgi`
module instead. The tox configuration has also been updated to allow
calling `django.setup` before the new module import.
