+++ storage_service/storage_service/wsgi.py
@@ -19,16 +19,23 @@ from os.path import abspath, dirname
 from sys import path
 
 from django.core.wsgi import get_wsgi_application
+import django
+
+django.setup()
+from common.startup import startup
 
 
 SITE_ROOT = dirname(dirname(abspath(__file__)))
 path.append(SITE_ROOT)
+
 
 # We defer to a DJANGO_SETTINGS_MODULE already in the environment. This breaks
 # if running multiple sites in the same mod_wsgi process. To fix this, use
 # mod_wsgi daemon mode with each site in its own daemon process, or use
 # os.environ["DJANGO_SETTINGS_MODULE"] = "jajaja.settings"
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "storage_service.settings.production")
+
+startup()
 
 # This application object is used by any WSGI server configured to use this
 # file. This includes Django's development server, if the WSGI_APPLICATION
