+++ tox.ini
@@ -48,5 +48,5 @@ setenv =
 exclude = .tox, .git, __pycache__, .cache, build, dist, *.pyc, *.egg-info, .eggs
 application-import-names = flake8
 select = C, E, F, W, B, B950
-ignore = E203, E501, W503
+ignore = E203, E402, E501, W503
 max-line-length = 80
